+++ fsic.py
@@ -5,7 +5,7 @@ fsic
 Tools for macroeconomic modelling in Python.
 """
 
-__version__ = '0.6.0.dev'
+__version__ = '0.6.1.dev'
 
 
 import copy